:zap: Performance and memory improvement (API)
- Now the board stores the last known location of both kings
- Switching from int to byte in Position struct

+++ src/back-end/FunChess/src/FunChess.Core.Chess/Extensions/CellArrayExtensions.cs
@@ -1,5 +1,3 @@
-using FunChess.Core.Chess.Constants;
-using FunChess.Core.Chess.Enums;
 using FunChess.Core.Chess.Structs;
 
 namespace FunChess.Core.Chess.Extensions;
@@ -8,7 +6,7 @@ public static class CellArrayExtensions
 {
     public static bool CheckNorth(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y + 1; y < move.Next.Y; y++)
+        for (byte y = (byte)(move.Previous.Y + 1); y < move.Next.Y; y++)
         {
             Position currentPosition = new(y, move.Previous.X);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -18,7 +16,7 @@ public static class CellArrayExtensions
 
     public static bool CheckSouth(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y - 1; y > move.Next.Y; y--)
+        for (byte y = (byte)(move.Previous.Y - 1); y > move.Next.Y; y--)
         {
             Position currentPosition = new(y, move.Previous.X);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -28,7 +26,7 @@ public static class CellArrayExtensions
 
     public static bool CheckEast(this Cell[] board, Move move)
     {
-        for (int x = move.Previous.X + 1; x < move.Next.X; x++)
+        for (byte x = (byte)(move.Previous.X + 1); x < move.Next.X; x++)
         {
             Position currentPosition = new(move.Previous.Y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -38,7 +36,7 @@ public static class CellArrayExtensions
 
     public static bool CheckWest(this Cell[] board, Move move)
     {
-        for (int x = move.Previous.X - 1; x > move.Next.X; x--)
+        for (byte x = (byte)(move.Previous.X - 1); x > move.Next.X; x--)
         {
             Position currentPosition = new(move.Previous.Y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -48,7 +46,7 @@ public static class CellArrayExtensions
 
     public static bool CheckNortheast(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y + 1, x = move.Previous.X + 1; y < move.Next.Y; y++, x++)
+        for (byte y = (byte)(move.Previous.Y + 1), x = (byte)(move.Previous.X + 1); y < move.Next.Y; y++, x++)
         {
             Position currentPosition = new(y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -58,7 +56,7 @@ public static class CellArrayExtensions
 
     public static bool CheckNorthwest(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y + 1, x = move.Previous.X - 1; y < move.Next.Y; y++, x--)
+        for (byte y = (byte)(move.Previous.Y + 1), x = (byte)(move.Previous.X - 1); y < move.Next.Y; y++, x--)
         {
             Position currentPosition = new(y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -68,7 +66,7 @@ public static class CellArrayExtensions
 
     public static bool CheckSoutheast(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y - 1, x = move.Previous.X + 1; y > move.Next.Y; y--, x++)
+        for (byte y = (byte)(move.Previous.Y - 1), x = (byte)(move.Previous.X + 1); y > move.Next.Y; y--, x++)
         {
             Position currentPosition = new(y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
@@ -78,7 +76,7 @@ public static class CellArrayExtensions
     
     public static bool CheckSouthwest(this Cell[] board, Move move)
     {
-        for (int y = move.Previous.Y - 1, x = move.Previous.X - 1; y > move.Next.Y; y--, x--)
+        for (byte y = (byte)(move.Previous.Y - 1), x = (byte)(move.Previous.X - 1); y > move.Next.Y; y--, x--)
         {
             Position currentPosition = new(y, x);
             if (!board[currentPosition.Index].IsEmpty()) return false;
