+++ src/back-end/FunChess/test/FunChess.Tests/Chess/Pieces/BishopTest.cs
@@ -12,9 +12,9 @@ public sealed class BishopTest
         Board board = new();
 
         // White bishops
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(new(0, 2), nextMove)));
@@ -23,9 +23,9 @@ public sealed class BishopTest
         }
         board.MovePiece(new Move(new Position(1, 0), new Position(2, 0)));
         // Black bishops
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(new(7, 2), nextMove)));
