+++ src/back-end/FunChess/test/FunChess.Tests/Chess/Pieces/KingTest.cs
@@ -15,9 +15,9 @@ public sealed class KingTest
         Position whiteKing = new(0, 3);
 
         // White king
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(whiteKing, nextMove)));
@@ -26,9 +26,9 @@ public sealed class KingTest
         board.MovePiece(new Move(new Position(1, 0), new Position(2, 0)));
 
         // Black king
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(blackKing, nextMove)));
