+++ src/back-end/FunChess/test/FunChess.Tests/Chess/Pieces/TowerTest.cs
@@ -12,9 +12,9 @@ public sealed class TowerTest
         Board board = new();
 
         // White towers
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(new(0, 0), nextMove)));
@@ -24,9 +24,9 @@ public sealed class TowerTest
         board.MovePiece(new Move(new Position(1, 0), new Position(2, 0)));
 
         // Black towers
-        for (int y = 0; y < BoardConstants.Length; y++)
+        for (byte y = 0; y < BoardConstants.Length; y++)
         {
-            for (int x = 0; x < BoardConstants.Length; x++)
+            for (byte x = 0; x < BoardConstants.Length; x++)
             {
                 Position nextMove = new(y, x);
                 Assert.False(board.PieceCanMove(new Move(new(7, 0), nextMove)));
